refactor: extract zero-runs writes from infra layer to zero platform layer
Move zeroRuns table writes and triggerSource/scheduleId fields from the
generic infra layer (run-service.ts, run-queue-service.ts) to the Zero
platform layer (zero-run-service.ts), since these are purely Zero
business concerns that the infra layer should not know about.

- Remove zeroRuns INSERT from createRun() and enqueueRun() transactions
- Add post-call zeroRuns INSERT in createZeroRun() after startRun() returns
- Move getRunQueueStatus() to new zero-queue-service.ts (it joins
  zeroRuns and zeroAgents, making it a Zero-layer concern)
- Remove triggerSource and scheduleId from CreateRunParams/StartRunParams
- Clean up unused imports in run-service.ts and run-queue-service.ts
- Update telegram test to mock createZeroRun instead of startRun

Closes #7154
