+++ turbo/apps/web/src/lib/run/__tests__/create-run.test.ts
@@ -15,7 +15,6 @@ import {
   createTestOrgModelProvider,
   insertStalePendingRun,
   findTestRunRecord,
-  findTestZeroRun,
   findTestRunCallbacks,
   findTestStorage,
   findTestRunnerJobEntry,
@@ -150,14 +149,6 @@ describe("createRun()", () => {
       const run = await findTestRunRecord(result.runId);
 
       expect(run!.secretNames).toEqual(["API_KEY", "DB_PASS"]);
-    });
-
-    it("should set null scheduleId when not provided", async () => {
-      const result = await createRun(baseParams());
-
-      const zeroRun = await findTestZeroRun(result.runId);
-
-      expect(zeroRun!.scheduleId).toBeNull();
     });
   });
