+++ turbo/apps/web/src/lib/zero/zero-queue-service.ts
@@ -1,0 +1,208 @@
+import {
+  eq,
+  and,
+  count,
+  gt,
+  or,
+  sql,
+  asc,
+  desc,
+  isNotNull,
+  avg,
+} from "drizzle-orm";
+import { agentRuns } from "../../db/schema/agent-run";
+import { zeroRuns } from "../../db/schema/zero-run";
+import {
+  agentComposeVersions,
+  agentComposes,
+} from "../../db/schema/agent-compose";
+import { zeroAgents } from "../../db/schema/zero-agent";
+import { getCachedUser } from "../auth/user-cache-service";
+import {
+  PENDING_RUN_TTL_MS,
+  getEffectiveConcurrencyLimit,
+} from "../run/run-service";
+import type { OrgTier, QueueResponse, TriggerSource } from "@vm0/core";
+
+const RECENT_RUNS_FOR_ETA = 20;
+const PROMPT_TRUNCATE_LENGTH = 200;
+
+/**
+ * Get run queue status for an org, including concurrency info,
+ * queued/running entries, and estimated time per run.
+ *
+ * Privacy filtering: non-owners see nullified personal fields.
+ *
+ * This is a Zero-layer concern because it joins zero_runs and zero_agents
+ * to enrich queue entries with triggerSource and agent display names.
+ */
+export async function getRunQueueStatus(
+  userId: string,
+  orgId: string,
+  orgTier: OrgTier,
+): Promise<QueueResponse> {
+  const db = globalThis.services.db;
+  const limit = getEffectiveConcurrencyLimit(orgTier);
+
+  // Count active runs (same logic as checkRunConcurrencyLimit)
+  const staleThreshold = new Date(Date.now() - PENDING_RUN_TTL_MS);
+  const [activeResult] = await db
+    .select({ count: count() })
+    .from(agentRuns)
+    .where(
+      and(
+        eq(agentRuns.orgId, orgId),
+        or(
+          eq(agentRuns.status, "running"),
+          and(
+            eq(agentRuns.status, "pending"),
+            gt(agentRuns.createdAt, staleThreshold),
+          ),
+        ),
+      ),
+    );
+  const active = Number(activeResult?.count ?? 0);
+
+  // Fetch queued runs in FIFO order (with extra fields for owner details)
+  const queuedRuns = await db
+    .select({
+      id: agentRuns.id,
+      runUserId: agentRuns.userId,
+      createdAt: agentRuns.createdAt,
+      agentName: agentComposes.name,
+      agentDisplayName: zeroAgents.displayName,
+      prompt: agentRuns.prompt,
+      triggerSource: zeroRuns.triggerSource,
+      continuedFromSessionId: agentRuns.continuedFromSessionId,
+    })
+    .from(agentRuns)
+    .leftJoin(zeroRuns, eq(agentRuns.id, zeroRuns.id))
+    .leftJoin(
+      agentComposeVersions,
+      eq(agentRuns.agentComposeVersionId, agentComposeVersions.id),
+    )
+    .leftJoin(
+      agentComposes,
+      eq(agentComposeVersions.composeId, agentComposes.id),
+    )
+    .leftJoin(zeroAgents, eq(agentComposes.id, zeroAgents.id))
+    .where(and(eq(agentRuns.orgId, orgId), eq(agentRuns.status, "queued")))
+    .orderBy(asc(agentRuns.createdAt));
+
+  // Fetch running tasks
+  const runningRuns = await db
+    .select({
+      id: agentRuns.id,
+      runUserId: agentRuns.userId,
+      startedAt: agentRuns.startedAt,
+      agentName: agentComposes.name,
+      agentDisplayName: zeroAgents.displayName,
+    })
+    .from(agentRuns)
+    .leftJoin(
+      agentComposeVersions,
+      eq(agentRuns.agentComposeVersionId, agentComposeVersions.id),
+    )
+    .leftJoin(
+      agentComposes,
+      eq(agentComposeVersions.composeId, agentComposes.id),
+    )
+    .leftJoin(zeroAgents, eq(agentComposes.id, zeroAgents.id))
+    .where(and(eq(agentRuns.orgId, orgId), eq(agentRuns.status, "running")))
+    .orderBy(asc(agentRuns.startedAt));
+
+  // Calculate estimated time per run from recent completed runs
+  const recentRuns = db
+    .select({
+      durationMs:
+        sql<number>`EXTRACT(EPOCH FROM (${agentRuns.completedAt} - ${agentRuns.startedAt})) * 1000`.as(
+          "duration_ms",
+        ),
+    })
+    .from(agentRuns)
+    .where(
+      and(
+        eq(agentRuns.orgId, orgId),
+        eq(agentRuns.status, "completed"),
+        isNotNull(agentRuns.completedAt),
+        isNotNull(agentRuns.startedAt),
+      ),
+    )
+    .orderBy(desc(agentRuns.completedAt))
+    .limit(RECENT_RUNS_FOR_ETA)
+    .as("recent_runs");
+  const [etaResult] = await db
+    .select({
+      avgMs: avg(recentRuns.durationMs),
+    })
+    .from(recentRuns);
+  const estimatedTimePerRun = etaResult?.avgMs
+    ? Math.round(Number(etaResult.avgMs))
+    : null;
+
+  // Resolve user emails in parallel (for both queued and running)
+  const allUserIds = [
+    ...new Set([
+      ...queuedRuns.map((r) => r.runUserId),
+      ...runningRuns.map((r) => r.runUserId),
+    ]),
+  ];
+  const userMap = new Map<string, string>();
+  await Promise.all(
+    allUserIds.map(async (uid) => {
+      const user = await getCachedUser(uid);
+      userMap.set(uid, user.email);
+    }),
+  );
+
+  // Build queue response with privacy filtering
+  const queue = queuedRuns.map((run, index) => {
+    const isOwner = run.runUserId === userId;
+    return {
+      position: index + 1,
+      agentName: isOwner ? (run.agentName ?? "unknown") : null,
+      agentDisplayName: isOwner ? (run.agentDisplayName ?? null) : null,
+      userEmail: isOwner ? (userMap.get(run.runUserId) ?? "unknown") : null,
+      createdAt: run.createdAt.toISOString(),
+      isOwner,
+      runId: isOwner ? run.id : null,
+      prompt: isOwner
+        ? run.prompt.length > PROMPT_TRUNCATE_LENGTH
+          ? run.prompt.slice(0, PROMPT_TRUNCATE_LENGTH) + "..."
+          : run.prompt
+        : null,
+      triggerSource: isOwner
+        ? ((run.triggerSource ?? "cli") as TriggerSource)
+        : null,
+      sessionLink:
+        isOwner && run.continuedFromSessionId
+          ? `/chat/${run.continuedFromSessionId}`
+          : null,
+    };
+  });
+
+  // Build running tasks response with privacy filtering
+  const runningTasks = runningRuns.map((run) => {
+    const isOwner = run.runUserId === userId;
+    return {
+      runId: isOwner ? run.id : null,
+      agentName: run.agentName ?? "unknown",
+      agentDisplayName: run.agentDisplayName ?? null,
+      userEmail: userMap.get(run.runUserId) ?? "unknown",
+      startedAt: run.startedAt?.toISOString() ?? null,
+      isOwner,
+    };
+  });
+
+  return {
+    concurrency: {
+      tier: orgTier,
+      limit,
+      active,
+      available: limit === 0 ? -1 : Math.max(0, limit - active),
+    },
+    queue,
+    runningTasks,
+    estimatedTimePerRun,
+  };
+}
